内核带颜色日志

+++ include/libc/string/str.h
@@ -118,7 +118,7 @@ finline int strncmp(cstr _s1, cstr _s2, size_t n) {
   while (s1 != e1) {
     c1 = *s1++;
     c2 = *s2++;
-    if (!c1) return c1 - c2;
+    if (!c1) return (int)c1 - (int)c2;
     if (c1 != c2) return c1 - c2;
   }
   return 0;
